getShowSurveyToast
